+++ utils/model.py
@@ -1,0 +1,6 @@
+users:list=[
+    {'name': 'Maciej', 'location':'Łódź', 'posts': 100},
+    {'name': 'Mateusz', 'location':'Poznań', 'posts': 200},
+    {'name': 'Maciej01', 'location':'Siedlce', 'posts': 300},
+    {'name': 'Konrad', 'location':'Grudziądz', 'posts': 400},
+]
